requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-groq
fastembed
faiss-cpu
pymupdf
python-dotenv
tiktoken
python-multipart